https://github.com/jackdewinter/pymarkdown/issues/605 - Refactored
