+++ publish/coverage.json
@@ -6,8 +6,8 @@
         "totalCovered": 3581
     },
     "lineLevel": {
-        "totalMeasured": 14533,
-        "totalCovered": 14533
+        "totalMeasured": 14563,
+        "totalCovered": 14563
     }
 }
