+++ publish/pylint_suppression.json
@@ -7,10 +7,17 @@
         },
         "pymarkdown/application_logging.py": {},
         "pymarkdown/bad_tokenization_error.py": {},
-        "pymarkdown/block_quote_data.py": {},
-        "pymarkdown/block_quote_processor.py": {
-            "too-many-arguments": 17,
+        "pymarkdown/block_quotes/block_quote_count_helper.py": {
+            "too-many-arguments": 7
+        },
+        "pymarkdown/block_quotes/block_quote_data.py": {},
+        "pymarkdown/block_quotes/block_quote_non_fenced_helper.py": {
+            "too-few-public-methods": 1,
+            "too-many-arguments": 6,
             "too-many-locals": 1
+        },
+        "pymarkdown/block_quotes/block_quote_processor.py": {
+            "too-many-arguments": 4
         },
         "pymarkdown/coalesce_processor.py": {
             "too-few-public-methods": 1
@@ -314,8 +321,8 @@
     },
     "disables-by-name": {
         "too-many-arguments": 172,
+        "too-few-public-methods": 22,
         "too-many-locals": 16,
-        "too-few-public-methods": 21,
         "chained-comparison": 1,
         "too-many-boolean-expressions": 1,
         "too-many-instance-attributes": 19,
