+++ pymarkdown/container_blocks/container_helper.py
@@ -8,7 +8,7 @@ container modules.
 import logging
 from typing import List, cast
 
-from pymarkdown.block_quote_data import BlockQuoteData
+from pymarkdown.block_quotes.block_quote_data import BlockQuoteData
 from pymarkdown.container_markdown_token import BlockQuoteMarkdownToken
 from pymarkdown.markdown_token import EndMarkdownToken, MarkdownToken
 from pymarkdown.parser_logger import ParserLogger
