+++ pymarkdown/leaf_blocks/setext_leaf_block_processor.py
@@ -4,7 +4,7 @@ Module to provide processing for the setext heading leaf blocks.
 import logging
 from typing import List, Optional, Tuple, cast
 
-from pymarkdown.block_quote_data import BlockQuoteData
+from pymarkdown.block_quotes.block_quote_data import BlockQuoteData
 from pymarkdown.leaf_markdown_token import (
     ParagraphMarkdownToken,
     SetextHeadingMarkdownToken,
